move keepalive types to own file
