+++ ydb/src/grpc_wrapper/raw_table_service/mod.rs
@@ -1,3 +1,4 @@
 pub(crate) mod client;
 pub(crate) mod create_session;
+pub(crate) mod keepalive;
 pub(crate) mod rollback_transaction;
